Fix issue with probe not turning off if it's the last operation
When probing is only needed once to set the WCS instead of every time a new part is loaded, it's a good idea to keep the probing operations at the bottom of the CAM setup, after a manual NC that will end the program. This way, when needed to probe, the operator can jump after the manual NC to run the probing ops. The benefit is that probing is kept together with the main program instead of a separate program, and the whole thing only needs to be post processed once as a whole.

The bug is a result of an earlier change, when probing is the last operation, no M81 will be generated, and the probe will be left on. The fix is to check to see when there's no next section, write M81.